[BED-23] Move CurrentItem to new package
Took 42 seconds
