+++ src/main/java/net/alphalightning/bedwars/game/ui/ArmorGui.java
@@ -1,7 +1,7 @@
 package net.alphalightning.bedwars.game.ui;
 
 import net.alphalightning.bedwars.game.ui.item.BuyableItem;
-import net.alphalightning.bedwars.game.ui.legacy.CurrentItem;
+import net.alphalightning.bedwars.game.ui.item.CurrentItem;
 import org.bukkit.Material;
 import xyz.xenondevs.invui.gui.Gui;
