+++ src/main/java/net/alphalightning/bedwars/game/ui/ExtrasGui.java
@@ -1,5 +1,6 @@
 package net.alphalightning.bedwars.game.ui;
 
+import net.alphalightning.bedwars.game.ui.item.CurrentItem;
 import net.alphalightning.bedwars.game.ui.legacy.*;
 import xyz.xenondevs.invui.gui.Gui;
