+++ src/main/java/net/alphalightning/bedwars/game/ui/FastBuyGui.java
@@ -1,6 +1,7 @@
 package net.alphalightning.bedwars.game.ui;
 
 import net.alphalightning.bedwars.game.ui.item.BuyableItem;
+import net.alphalightning.bedwars.game.ui.item.CurrentItem;
 import net.alphalightning.bedwars.game.ui.legacy.*;
 import org.bukkit.Material;
 import xyz.xenondevs.invui.gui.Gui;
